+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_compose_command.py
@@ -2,7 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-
+import os
 import unittest  # pylint: disable=unused-import
 
 from azure.cli.testsdk import (ResourceGroupPreparer)
@@ -19,6 +19,11 @@ class ContainerappComposePreviewCommandScenarioTest(ContainerappComposePreviewSc
     @serial_test()
     @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
     def test_containerapp_compose_with_command_string(self, resource_group):
+        location = os.getenv("CLITestLocation")
+        if not location:
+            location = 'eastus'
+        self.cmd('configure --defaults location={}'.format(location))
+
         compose_text = """
 services:
   foo:
@@ -37,7 +42,7 @@ services:
             'compose': compose_file_name,
         })
 
-        create_containerapp_env(self, env_name, resource_group, 'eastus')
+        create_containerapp_env(self, env_name, resource_group)
 
         command_string = 'containerapp compose create'
         command_string += ' --compose-file-path {compose}'
@@ -52,6 +57,11 @@ services:
     @serial_test()
     @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
     def test_containerapp_compose_with_command_list(self, resource_group):
+        location = os.getenv("CLITestLocation")
+        if not location:
+            location = 'eastus'
+        self.cmd('configure --defaults location={}'.format(location))
+
         compose_text = """
 services:
   foo:
@@ -68,7 +78,7 @@ services:
             'compose': compose_file_name,
         })
 
-        create_containerapp_env(self, env_name, resource_group, 'eastus')
+        create_containerapp_env(self, env_name, resource_group)
         
         command_string = 'containerapp compose create'
         command_string += ' --compose-file-path {compose}'
@@ -83,6 +93,11 @@ services:
     @serial_test()
     @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
     def test_containerapp_compose_with_command_list_and_entrypoint(self, resource_group):
+        location = os.getenv("CLITestLocation")
+        if not location:
+            location = 'eastus'
+        self.cmd('configure --defaults location={}'.format(location))
+
         compose_text = """
 services:
   foo:
@@ -101,7 +116,7 @@ services:
             'compose': compose_file_name,
         })
 
-        create_containerapp_env(self, env_name, resource_group, 'eastus')
+        create_containerapp_env(self, env_name, resource_group)
 
         command_string = 'containerapp compose create'
         command_string += ' --compose-file-path {compose}'