+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_compose_registries.py
@@ -2,7 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-
+import os
 import unittest  # pylint: disable=unused-import
 
 from azure.cli.testsdk import (ResourceGroupPreparer)
@@ -18,6 +18,11 @@ class ContainerappComposePreviewRegistryAllArgsScenarioTest(ContainerappComposeP
     @serial_test()
     @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
     def test_containerapp_compose_create_with_registry_all_args(self, resource_group):
+        location = os.getenv("CLITestLocation")
+        if not location:
+            location = 'eastus'
+        self.cmd('configure --defaults location={}'.format(location))
+
         compose_text = """
 services:
   foo:
@@ -59,6 +64,11 @@ class ContainerappComposePreviewRegistryServerArgOnlyScenarioTest(ContainerappCo
     @serial_test()
     @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
     def test_containerapp_compose_create_with_registry_server_arg_only(self, resource_group):
+        location = os.getenv("CLITestLocation")
+        if not location:
+            location = 'eastus'
+        self.cmd('configure --defaults location={}'.format(location))
+
         compose_text = """
 services:
   foo: